Move model class to modeltypes module
